Move unit test to its own file with the same package location as BatchOrderApiResponse
