+++ example/src/test/java/org/wordpress/android/fluxc/wc/order/OrderEntityTest.kt
@@ -145,28 +145,4 @@ class OrderEntityTest {
         assertEquals("Flat Rate Shipping", shippingLinesList[0].methodTitle)
         assertEquals("Local Pickup Shipping", shippingLinesList[1].methodTitle)
     }
-
-    @Test
-    fun testDeserializeBatchOrderResponse() {
-        val testGson = GsonBuilder()
-            .create()
-
-        val batchOrderJson = UnitTestUtils.getStringFromResourceFile(
-            this.javaClass, "wc/orders-batch.json"
-        )
-
-        val response = testGson.fromJson(batchOrderJson, BatchOrderApiResponse::class.java)
-        val orders = response.update
-
-        assertEquals(2, orders.size)
-
-        val firstOrder = orders[0] as BatchOrderApiResponse.OrderResponse.Success
-        assertEquals(1032L, firstOrder.order.id)
-        assertEquals("224.00", firstOrder.order.total)
-
-        val secondOrder = orders[1] as BatchOrderApiResponse.OrderResponse.Error
-        assertEquals(525L, secondOrder.id)
-        assertEquals("woocommerce_rest_shop_order_invalid_id", secondOrder.error.code)
-        assertEquals(400, secondOrder.error.data.status)
-    }
 }
